Improve the DefaultServlet so it can be properly subclassed.
